# webapp/models/swagger-spec.yaml
{"openapi":"3.0.3","info":{"title":"Experts","version":"1.39.0-dev-36-gfbf935e-dirty","description":"The Experts API specifies updates to a particular expert. Publically available API endpoints can be used for access to an experts data.  The permissions of current user allow additional access to the data.","termsOfService":"http://swagger.io/terms/","contact":{"email":"experts@ucdavis.edu"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"components":{"parameters":{"expertId":{"name":"expertId","in":"path","required":true,"schema":{"type":"string","format":"nano(\\d{8})","description":"The unique identifier for the expert"}},"relationshipId":{"name":"relationshipId","in":"path","required":true,"schema":{"type":"string","format":"urlencoded","description":"A unique identifier for an expert relationship"}},"p":{"in":"query","name":"p","description":"The letter the experts last name starts with","required":false,"schema":{"type":"string"}},"page":{"in":"query","name":"page","description":"The pagination of results to return, defaults to 1","required":false,"schema":{"type":"integer"}},"size":{"in":"query","name":"size","description":"The number of results to return per page, defaults to 25","required":false,"schema":{"type":"integer"}}},"schemas":{"Expert":{"type":"object","properties":{"@id":{"type":"string","description":"The unique identifier for the expert."},"@type":{"type":"array","items":{"type":"string"},"description":"The type of the expert."},"rank":{"type":"integer","description":"The rank of the expert."},"name":{"type":"string","description":"The name of the expert."},"url":{"type":"string","format":"url","description":"The URL related to the expert."},"hasEmail":{"type":"string","format":"email","description":"The email address of the expert."},"hasName":{"type":"object","properties":{"@id":{"type":"string","description":"The unique identifier for the name."},"@type":{"type":"string","description":"The type of the name."},"family":{"type":"string","description":"The family name of the expert."},"given":{"type":"string","description":"The given name of the expert."},"pronouns":{"type":"string","description":"The pronouns of the expert."}},"required":["@id","@type","family","given","pronouns"]},"hasTitle":{"type":"object","properties":{"@id":{"type":"string","description":"The unique identifier for the title."},"@type":{"type":"string","description":"The type of the title."},"name":{"type":"string","description":"The title of the expert."}},"required":["@id","@type","name"]},"hasOrganizationalUnit":{"type":"object","properties":{"@id":{"type":"string","description":"The unique identifier for the organizational unit."},"name":{"type":"string","description":"The name of the organizational unit."}},"required":["@id","name"]},"roles":{"type":"array","items":{"type":"string"},"description":"The roles of the expert."}},"required":["@id","@type","rank","name","url","hasEmail","hasName","hasTitle","hasOrganizationalUnit","roles"]},"Relationship":{"type":"object","properties":{"@id":{"type":"string"},"@type":{"type":"array","items":{"type":"string"}},"@graph":{"type":"array","items":{"type":"object","properties":{"@id":{"type":"string"},"@type":{"type":"array","items":{"type":"string"}},"is-visible":{"type":"boolean"},"rank":{"type":"integer","format":"int32"},"relates":{"type":"array","items":{"type":"string"}},"_":{"type":"object","properties":{"event":{"type":"object","properties":{"id":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"updateType":{"type":"array","items":{"type":"string"}}},"required":["id","timestamp","updateType"]},"updated":{"type":"string","format":"date-time"}},"required":["event","updated"]}},"required":["@id","@type","is-visible","rank","relates","_"]}},"roles":{"type":"array","items":{"type":"string"}}},"required":["@id","@type","@graph","roles"]},"Browse":{"type":"object","properties":{"total":{"type":"integer"},"hits":{"type":"array","items":{"type":"object","properties":{"contactInfo":{"type":"object","properties":{"hasURL":{"type":"array","items":{"type":"object","properties":{"@type":{"type":"array","items":{"type":"string"}},"@id":{"type":"string"},"url":{"type":"string"},"name":{"type":"string"},"rank":{"type":"integer"}}}},"hasEmail":{"type":"string"},"hasName":{"type":"object","properties":{"given":{"type":"string"},"@type":{"type":"string"},"pronouns":{"type":"string"},"@id":{"type":"string"},"family":{"type":"string"}}},"name":{"type":"string"},"hasTitle":{"type":"object","properties":{"@type":{"type":"string"},"name":{"type":"string"},"@id":{"type":"string"}}},"hasOrganizationalUnit":{"type":"object","properties":{"name":{"type":"string"},"@id":{"type":"string"}}}}},"name":{"type":"string"},"@id":{"type":"string"}}}}}},"Search":{"type":"object","properties":{"total":{"type":"integer"},"hits":{"type":"array","items":{"type":"object","properties":{"contactInfo":{"type":"object","properties":{"hasEmail":{"type":"string"},"hasName":{"type":"object","properties":{"given":{"type":"string"},"@type":{"type":"string"},"@id":{"type":"string"},"family":{"type":"string"}}},"name":{"type":"string"},"hasTitle":{"type":"object","properties":{"@type":{"type":"string"},"name":{"type":"string"},"@id":{"type":"string"}}},"hasOrganizationalUnit":{"type":"object","properties":{"name":{"type":"string"},"@id":{"type":"string"}}}}},"@type":{"type":"string"},"name":{"type":"string"},"@id":{"type":"string"},"_inner_hits":{"type":"array","items":{"type":"object","properties":{"volume":{"type":"string"},"@type":{"type":"array","items":{"type":"string"}},"author":{"type":"array","items":{"type":"object","properties":{"given":{"type":"string"},"rank":{"type":"integer"},"@id":{"type":"string"},"family":{"type":"string"}}}},"container-title":{"type":"string"},"ISSN":{"type":"string"},"abstract":{"type":"string"},"page":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"issued":{"type":"string"},"status":{"type":"string"}}}}}}}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},"responses":{"Expert":{"description":"The expert","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Expert"}}}},"Expert_not_found":{"description":"Expert not found"},"Expert_deleted":{"description":"Expert deleted"},"No_content":{"description":"No Content"},"Relationship":{"description":"The relationship","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Relationship"}}}},"Relationship_not_found":{"description":"Relationship not found"},"Browse":{"description":"The list of experts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Browse"}}}},"Invalid_request":{"description":"Invalid request"},"Search":{"description":"The list of search results","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Search"}}}},"Successful_operation":{"description":"Successful operation"},"Invalid_ID_supplied":{"description":"Invalid ID supplied"}},"requestBodies":{"Relationship_patch":{"content":{"application/json":{"schema":{"type":"object","properties":{"@id":{"type":"string"},"visible":{"type":"boolean"},"grant":{"type":"boolean"}}}}}},"Expert_patch":{"content":{"application/json":{"schema":{"type":"object","properties":{"@id":{"type":"string"},"visible":{"type":"boolean"}}}}}}}},"servers":[{"url":"http://localhost/api/expert"}],"tags":[{"name":"expert","description":"Expert Information"}],"paths":{"/api/browse/":{"get":{"description":"Returns counts for experts A - Z, or if sending query param p={letter}, will return results for experts with last names of that letter","parameters":[{"$ref":"#/components/parameters/p"},{"$ref":"#/components/parameters/page"},{"$ref":"#/components/parameters/size"}],"responses":{"200":{"$ref":"#/components/responses/Browse"},"400":{"$ref":"#/components/responses/Invalid_request"}}}},"/api/search/":{"get":{"description":"Returns matching search results for experts, including the number of matching works and grants","parameters":[{"$ref":"#/components/parameters/p"},{"$ref":"#/components/parameters/page"},{"$ref":"#/components/parameters/size"}],"responses":{"200":{"$ref":"#/components/responses/Search"},"400":{"$ref":"#/components/responses/Invalid_request"}}}},"/api/expert/{expertId}/{relationshipId}":{"get":{"description":"Get an expert relationship by id","parameters":[{"name":"expertId","in":"path","required":true,"schema":{"type":"string"}},{"name":"relationshipId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/Relationship"},"404":{"$ref":"#/components/responses/Relationship_not_found"}}},"patch":{"description":"Update an expert relationship by id","parameters":[{"name":"expertId","in":"path","required":true,"schema":{"type":"string"}},{"name":"relationshipId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/Relationship_patch"},"responses":{"204":{"$ref":"#/components/responses/No_content"},"404":{"$ref":"#/components/responses/Relationship_not_found"}}},"delete":{"description":"Update an expert relationship by id","parameters":[{"name":"expertId","in":"path","required":true,"schema":{"type":"string"}},{"name":"relationshipId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"$ref":"#/components/responses/No_content"},"404":{"$ref":"#/components/responses/Relationship_not_found"}}}},"/api/expert/{expertId}":{"get":{"description":"Get an expert by id","parameters":[{"name":"expertId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"$ref":"#/components/responses/Expert"},"404":{"$ref":"#/components/responses/Expert_not_found"}}},"patch":{"description":"Update an experts visibility by expert id","parameters":[{"name":"expertId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"$ref":"#/components/requestBodies/Expert_patch"},"responses":{"204":{"$ref":"#/components/responses/No_content"}}},"delete":{"description":"Delete an expert by id","parameters":[{"name":"expertId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"$ref":"#/components/responses/Expert_deleted"}}}},"/api/sitefarm/experts/{ids}":{"get":{"description":"Returns a JSON array of expert profiles","parameters":[{"name":"ids","in":"path","required":true,"schema":{"type":"string"},"description":"A comma separated list of expert IDs"}],"responses":{"200":{"$ref":"#/components/responses/Successful_operation"},"400":{"$ref":"#/components/responses/Invalid_ID_supplied"},"404":{"$ref":"#/components/responses/Expert_not_found"}}}}}}